+++ api/models/__init__.py
@@ -1,5 +1,8 @@
+from .client import Client, ClientServiceAuthorization
 from .task import Task
 
 __all__ = [
+    "Client",
+    "ClientServiceAuthorization",
     "Task",
 ]