+++ api/schemas/__init__.py
@@ -2,6 +2,7 @@ from .callback import TaskCallback
 from .queue import QueueData, QueueTask
 from .service import ServiceInfo
 from .status import HealthResponse, ReadyComponent, ReadyResponse
+from .storage import StorageUploadResponse
 from .task import (
     Callback,
     TaskDataFailed,
@@ -13,23 +14,55 @@ from .task import (
     TaskRequest,
     TaskResponse,
 )
+from .usage import ClientUsageResponse, ServiceUsage
 
-__all__ = [
-    "Callback",
-    "HealthResponse",
+_callback_schemas = [
+    "TaskCallback",
+]
+
+_queue_schemas = [
     "QueueData",
     "QueueTask",
+]
+
+_service_schemas = [
+    "ServiceInfo",
+]
+
+_status_schemas = [
+    "HealthResponse",
     "ReadyComponent",
     "ReadyResponse",
-    "ServiceInfo",
-    "TaskCallback",
+]
+
+_storage_schemas = [
+    "StorageUploadResponse",
+]
+
+_task_schemas = [
+    "Callback",
     "TaskDataFailed",
     "TaskDataPending",
     "TaskDataProgress",
     "TaskDataSuccess",
     "TaskErrorResponse",
     "TaskInfo",
-    "TaskInfo",
     "TaskRequest",
     "TaskResponse",
+]
+
+# usage
+_usage_schemas = [
+    "ClientUsageResponse",
+    "ServiceUsage",
+]
+
+__all__ = [
+    *_callback_schemas,
+    *_queue_schemas,
+    *_service_schemas,
+    *_status_schemas,
+    *_storage_schemas,
+    *_task_schemas,
+    *_usage_schemas,
 ]